Backend_API/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
pytest
httpx
